move app pkg validation funcs into dedicated file
